+++ packages/sampler/stories/moonstone-stories/Group.js
@@ -6,8 +6,10 @@ import SwitchItem from '@enact/moonstone/SwitchItem';
 import ToggleButton from '@enact/moonstone/ToggleButton';
 import Group from '@enact/ui/Group';
 import React from 'react';
-import {storiesOf, action} from '@kadira/storybook';
-import {boolean, select} from '@kadira/storybook-addon-knobs';
+import {storiesOf} from '@storybook/react';
+import {action} from '@storybook/addon-actions';
+import {boolean, select} from '@storybook/addon-knobs';
+import {withInfo} from '@storybook/addon-info';
 
 // Set up some defaults for info and knobs
 const prop = {
@@ -22,11 +24,10 @@ const prop = {
 
 const getComponent = (name) => prop.children[name];
 
-storiesOf('Group')
-	.addWithInfo(
-		' ',
-		'Basic usage of Group',
-		() => (
+storiesOf('UI', module)
+	.add(
+		'Group',
+		withInfo('Basic usage of Group')(() => (
 			<Group
 				childComponent={getComponent(select('childComponent', Object.keys(prop.children), 'CheckboxItem'))}
 				itemProps={{
@@ -39,5 +40,5 @@ storiesOf('Group')
 			>
 				{['Item 1', 'Item 2', 'Item 3']}
 			</Group>
-		)
+		))
 	);