+++ packages/sampler/stories/qa-stories/Remeasurable.js
@@ -3,7 +3,7 @@ import Item from '@enact/moonstone/Item';
 import Scroller from '@enact/moonstone/Scroller';
 import ri from '@enact/ui/resolution';
 import React from 'react';
-import {storiesOf} from '@kadira/storybook';
+import {storiesOf} from '@storybook/react';
 import SelectableItem from '@enact/moonstone/SelectableItem';
 
 const data = [
@@ -24,8 +24,8 @@ class NoUpdate extends React.Component {
 	}
 }
 
-storiesOf('Remeasurable')
-	.addWithInfo(
+storiesOf('Remeasurable', module)
+	.add(
 		'should recalculate long marquee when scrollbar is rendered',
 		() => (
 			<Scroller style={{height: ri.unit(399, 'rem'), width: ri.unit(501, 'rem')}}>
@@ -43,7 +43,7 @@ storiesOf('Remeasurable')
 			</Scroller>
 		)
 	)
-	.addWithInfo(
+	.add(
 		'should recalculate when selectable item is selected',
 		() => (
 			<Scroller style={{height: ri.unit(399, 'rem'), width: ri.unit(501, 'rem')}}>