+++ packages/moonstone/CHANGELOG.md
@@ -11,10 +11,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Changed
 
 - `moonstone/Marquee` to do less-costly calculations during measurement and optimized the applied styles
-- `moonstone/MoonstoneDecorator` root node to fill the entire space available, which simplifies positioning and sizing for child elements (previously always measured 0 in height)
 
 ### Fixed
 
+- `moonstone/VirtualList` to render properly with fiber reconciler
 - `moonstone/VirtualList` to work focus option in scrollTo api
 - `moonstone/ExpandableSpotlightDecorator` to not spot the title upon collapse when in `pointerMode`